Start restructuring the cluster repos.
Each part of the EDO needs to be able to answer the question 'which of
these manifests are mine'. Without the K8s API to search for labels this
means using the repo directory structure.

Instead of an NS/KIND/NAME.yaml directory structure, organise the
manifests by owner and subsystem. So EDO manifests go under `edo`,
manually added manifests go under `by-hand`, and so on. As long as this
is kept to questions like 'where are all the edge agent deployment
manifests' can be answered fairly straightforwardly.

Start moving the manifest-handling functions out of Checkout and into a
separate file. Rename the manifest templates to `templates.js`.
