+++ lib/checkout.js
@@ -16,7 +16,7 @@ import yaml from "yaml";
 import { Debug, UUIDs } from "@amrc-factoryplus/utilities";
 
 import { Git } from "./uuids.js";
-import * as manifests from "./manifests.js";
+import * as tmpl from "./templates.js";
 
 const debug = new Debug();
 
@@ -203,5 +203,13 @@ export class Checkout {
         return files
             .filter(f => f.endsWith(".yaml"))
             .map(f => [namespace, kind, f.slice(0, -5)]);
+    }
+
+    /* Return an absolute path for manifest files managed by a
+     * particular subsystem (e.g. sealed secrets). Currently the cluster
+     * is ignored but it might be used in future with a multi-cluster
+     * repo structure. */
+    path_for (cluster, subsys, ...paths) {
+        return path.join(this.dir, "edo", subsys, ...paths);
     }
 }
