+++ lib/clusters.js
@@ -11,8 +11,8 @@ import { Debug, UUIDs }         from "@amrc-factoryplus/utilities";
 import { Checkout }             from "./checkout.js";
 import { KrbKeys }              from "./krbkeys.js";
 import { write_sealed_secret }  from "./kubeseal.js";
-import * as manifests           from "./manifests.js";
-import { FLUX }                 from "./manifests.js";
+import * as tmpl                from "./templates.js";
+import { FLUX }                 from "./templates.js";
 import { Git, Edge }            from "./uuids.js";
 
 const debug = new Debug();
@@ -179,7 +179,7 @@ export class Clusters {
             fplus:  this.fplus, 
             url:    spec.flux,
         });
-        await co.write_file("README.md", manifests.README);
+        await co.write_file("README.md", tmpl.README);
         await co.commit("Add README.");
         await this.write_username_secret(co, spec);
         await this.setup_repo_links(co, spec);
@@ -210,8 +210,8 @@ export class Clusters {
         const url = new URL(`git/${repo}`, git_base).toString();
 
         this.log("Adding source %s", url);
-        await co.write_manifest(manifests.git_repo(name, url, spec));
-        await co.write_manifest(manifests.flux_kust(name, spec));
+        await co.write_manifest(tmpl.git_repo(name, url, spec));
+        await co.write_manifest(tmpl.flux_kust(name, spec));
     }
 
     async setup_repo_links (co, spec) {
@@ -229,7 +229,7 @@ export class Clusters {
     async create_namespace (co, spec) {
         const { namespace } = spec.template.cluster;
 
-        await co.write_manifest(manifests.namespace(namespace));
+        await co.write_manifest(tmpl.namespace(namespace));
         await co.commit("Written namespace manifest.");
     }
