+++ lib/krbkeys.js
@@ -8,7 +8,7 @@ import k8s          from "@kubernetes/client-node";
 
 import { Debug }    from "@amrc-factoryplus/utilities";
 
-import { KRB, FLUX, krb_key }  from "./manifests.js";
+import { KRB, FLUX, krb_key }  from "./templates.js";
 
 const debug = new Debug();
