+++ lib/kubeseal.js
@@ -10,7 +10,7 @@ import fs_p from "fs/promises";
 import concat_stream from "concat-stream";
 import tmpfile from "tmp-promise";
 
-import * as manifests from "./manifests.js";
+import * as tmpl from "./templates.js";
 
 export async function write_sealed_secret (opts) {
     const { checkout: co, namespace, name, key } = opts;
@@ -19,7 +19,7 @@ export async function write_sealed_secret (opts) {
     opts.log("Got sealed data: %s", sealed);
 
     const obj = await co.read_manifest(namespace, "SealedSecret", name) ??
-        manifests.sealed_secret(namespace, name);
+        tmpl.sealed_secret(namespace, name);
     obj.spec.encryptedData[key] = sealed;
     await co.write_manifest(obj);
     await co.commit("Updated sealed secret %s/%s/%s.",
