+++ lib/reconciler.js
@@ -11,7 +11,7 @@ import { Debug, UUIDs } from "@amrc-factoryplus/utilities";
 
 import { Checkout }         from "./checkout.js";
 import { Clusters }         from "./clusters.js";
-import { flux_helm }        from "./manifests.js";
+import { flux_helm }        from "./templates.js";
 import { SealedSecrets }    from "./secrets.js";
 import { Edge }             from "./uuids.js";
